lume-agent/requirements.txt
# Lume AI Productivity Agent - Python Dependencies

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for OAuth requests
httpx==0.25.2

# Database and ORM
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Authentication and security
pyjwt==2.8.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Google API clients (for future integration)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Gemini AI SDK
google-generativeai==0.3.1

# Additional utilities
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
